Add files via upload
In This ALPHA-NUMERIC CHARACTER RECOGNITION model, we are going to recognize
digits and alphabets from their images. We take around 150 images of each alphabet and digit
in our training dataset. For classification we are going to use two classification algorithms, MLP
and CNN. Our Result shows that MLP is overfitting the training data. And thus the testing
accuracy of CNN algorithm is better than MLP algorithm